+++ lunavoxel/server/Reducers/ModifyBlockReducer.cs
@@ -3,7 +3,7 @@ using SpacetimeDB;
 public static partial class Module
 {
     [Reducer]
-    public static void ModifyBlock(ReducerContext ctx, string projectId, uint[] diffData, int layerIndex)
+    public static void ModifyBlock(ReducerContext ctx, string projectId, byte[] diffData, int layerIndex)
     {
         EnsureAccessToProject.Check(ctx, projectId, ctx.Sender);
 
@@ -12,7 +12,7 @@ public static partial class Module
 
         if (layer.Locked) return;
 
-        uint[] voxels = VoxelCompression.Decompress(layer.Voxels);
+        byte[] voxels = VoxelCompression.Decompress(layer.Voxels);
         for (int i = 0; i < diffData.Length; i++)
         {
             if (diffData[i] != 0)
